add Cronjob weekly&monthly, edit .sh

+++ data/cron/monthly-backup.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
+echo "Start monthly GitLab backup"
 wbpath=/nfs/backups/weekly/
 mbpath=/nfs/backups/monthly/
 wblast="$(sudo ls -ltr $wbpath | tail -1 | awk '{print $9}')"
 datenow="$(date +%d-%m-%y)"
-fullpath=$wbpath$wblast
-sudo mv $fullpath $mbpath
-echo "Move monthly backup path: $mbpath$wblast"
+fullbpath=$wbpath$wblast
+sudo cp $fullbpath $mbpath
+echo "Copy monthly backup path: $mbpath$wblast"
 allwb="$(find $wbpath -type f -print0 | xargs -0)"
-sudo rm $allwb
-echo "Weekly backup clear: $allwb"
+
+if [ -f "$mbpath$wblast" ]; then
+	sudo rm $allwb
+	echo "Weekly backup remove: $allwb"
+else
+    echo "Monthly Backup failed! $mbpath$wblast does not exist"
+fi
